+++ heksher/clients/async_client.py
@@ -12,7 +12,7 @@ from ordered_set import OrderedSet
 
 from heksher.clients.subclasses import AsyncContextManagerMixin, ContextFeaturesMixin, V1APIClient
 from heksher.clients.util import SettingsOutput
-from heksher.setting import MISSING, Setting
+from heksher.setting import NO_DEFAULT, Setting
 
 logger = getLogger(__name__)
 
@@ -73,7 +73,7 @@ class AsyncHeksherClient(V1APIClient, ContextFeaturesMixin, AsyncContextManagerM
                 'type': setting.type.heksher_string(),
                 'metadata': setting.metadata,
             }
-            if setting.default_value is not MISSING:
+            if setting.default_value is not NO_DEFAULT:
                 declaration_data['default_value'] = setting.default_value
             response = await self._http_client.put('api/v1/settings/declare', content=orjson.dumps(declaration_data),
                                                    headers=content_header)