+++ tests/unittest/test_settings.py
@@ -1,6 +1,6 @@
 import gc
 from logging import WARNING
-from typing import Sequence
+from typing import Sequence, Tuple
 from weakref import ref
 
 from pytest import raises
@@ -123,16 +123,18 @@ def test_useless_vals():
 def test_setting_callback():
     a = Setting('a', int, 'abcx', default_value=-1)
 
-    def setting_callback_1(setting: Setting, rule: Sequence[str], value: int) -> int:
+    def setting_callback_1(value: int, rule: Sequence[Tuple[str, str]], setting: Setting) -> int:
         if setting.name == 'a' and value == 10:
             return 12
+        return value
 
-    def setting_callback_2(setting: Setting, rule: Sequence[str], value: int) -> int:
+    def setting_callback_2(value: int, rule: Sequence[Tuple[str, str]], setting: Setting) -> int:
         if setting.name == 'a' and value == 12:
             return 7
+        return value
 
-    a.add_callback(setting_callback_1)
-    a.add_callback(setting_callback_2)
+    a.add_validator(setting_callback_1)
+    a.add_validator(setting_callback_2)
     c1 = SyncStubHeksherClient()
     c1.set_as_main()
     c1.patch(a, 10)
@@ -142,16 +144,18 @@ def test_setting_callback():
 def test_setting_rules_collection_callback():
     a = Setting('a', int, 'abcx', default_value=-1)
 
-    def setting_callback_1(setting: Setting, rule: Sequence[str], value: int) -> int:
+    def setting_callback_1(value: int, rule: Sequence[Tuple[str, str]], setting: Setting) -> int:
         if setting.name == 'a' and value == 0:
             return 3
+        return value
 
-    def setting_callback_2(setting: Setting, rule: Sequence[str], value: int) -> int:
+    def setting_callback_2(value: int, rule: Sequence[Tuple[str, str]], setting: Setting) -> int:
         if setting.name == 'a' and value == 3:
             return 7
+        return value
 
-    a.add_callback(setting_callback_1)
-    a.add_callback(setting_callback_2)
+    a.add_validator(setting_callback_1)
+    a.add_validator(setting_callback_2)
     c1 = SyncStubHeksherClient()
     c1.set_as_main()